market-prices: refactor ft_fund to search for symbols, not ISIN
Some funds cannot be queried with their ISIN, or don't have the currency
in their ticker symbol (maybe because they are only available in one
currency). Adapt our internal API in order to query those funds. Also be
backwards compatible with user's existing commodities.json and use the
ISIN when it is specified.

Note that this breaks the use case where the ISIN is specified as the
commodity name because then the currency will be missing from the symbol
being queried, but I see no other good way to migrate this case.

Also add new usage examples to the example commodities.json, so we can
use it as a integration test.